Refactored some functions into a utility class - it's only the beginning.
