+++ External/Plugins/ASCompletion/ASCompletion.csproj
@@ -77,6 +77,7 @@
     <Compile Include="Completion\ArgumentsProcessor.cs" />
     <Compile Include="Completion\ASComplete.cs" />
     <Compile Include="Completion\ASGenerator.cs" />
+    <Compile Include="Completion\CodeUtils.cs" />
     <Compile Include="Completion\Reformater.cs" />
     <Compile Include="Completion\TemplateUtils.cs" />
     <Compile Include="Context\ASContext.cs" />
