+++ External/Plugins/ASCompletion/Completion/ASComplete.cs
@@ -1229,13 +1229,13 @@ namespace ASCompletion.Completion
             while (tempLine > 0)
             {
                 tempText = Sci.GetLine(tempLine).Trim();
-                if (insideClass && IsTypeDecl(tempText, features.typesKeywords))
+                if (insideClass && CodeUtils.IsTypeDecl(tempText, features.typesKeywords))
                 {
                     tempIndent = Sci.GetLineIndentation(tempLine);
                     tab = tempIndent + Sci.TabWidth;
                     break;
                 }
-                if (tempText.Length > 0 && (tempText.EndsWith("}") || IsDeclaration(tempText, features)))
+                if (tempText.Length > 0 && (tempText.EndsWith("}") || CodeUtils.IsDeclaration(tempText, features)))
                 {
                     tempIndent = Sci.GetLineIndentation(tempLine);
                     tab = tempIndent;
@@ -1258,45 +1258,6 @@ namespace ASCompletion.Completion
             // show
             CompletionList.Show(known, autoHide, tail);
             return true;
-        }
-
-        /// <summary>
-        /// Lookup type declaration keywords anywhere in the provided text
-        /// </summary>
-        private static bool IsTypeDecl(string line, string[] typesKeywords)
-        {
-            var max = line.Length - 1;
-            foreach (string keyword in typesKeywords)
-            {
-                var p = line.IndexOf(keyword);
-                if (p >= 0) 
-                {
-                    // verify keyword between spaces
-                    var end = p + keyword.Length;
-                    if ((p == 0 || line[p-1] <= 32) 
-                        && end < max && line[end] <= 32) return true;
-                }
-            }
-            return false;
-        }
-
-        /// <summary>
-        /// Look if the provided text starts with any declaration keyword
-        /// </summary>
-        private static bool IsDeclaration(string line, ContextFeatures features)
-        {
-            foreach (string keyword in features.accessKeywords)
-                if (line.StartsWith(keyword) && SpaceFollows(line, keyword)) return true;
-            foreach (string keyword in features.declKeywords)
-                if (line.StartsWith(keyword) && SpaceFollows(line, keyword)) return true;
-            return false;
-        }
-
-        private static bool SpaceFollows(string line, string keyword)
-        {
-            var len = keyword.Length;
-            if (line.Length > len) return line[len] <= 32;
-            else return true;
         }
 
         #endregion
